actorwithbody splitted in favour of new particle type
Former-commit-id: 121bd069d0a9eeef60f5ecb085a11a93c4b4a84d
Former-commit-id: 539b4b6916e808c01298190cf347e928f61fe62e

+++ src/net/torvald/terrarum/mapgenerator/NoiseFilterCubic.kt
@@ -30,7 +30,7 @@ import com.jme3.math.FastMath
  * Created by minjaesong on 16-03-31.
  */
 object NoiseFilterCubic : NoiseFilter {
-    override fun getGrad(func_argX: Int, start: Float, end: Float): Float {
+    override fun getGrad(func_argX: Int, start: Double, end: Double): Double {
         val graph_gradient = -FastMath.pow(FastMath.pow((1 - WorldGenerator.TERRAIN_AVERAGE_HEIGHT).toFloat(), 3f), -1f) * // 1/4 -> 3/4 -> 9/16 -> 16/9
                              (start - end) / FastMath.pow(WorldGenerator.HEIGHT.toFloat(), 3f) *
                              FastMath.pow((func_argX - WorldGenerator.HEIGHT).toFloat(), 3f) + end
